Multivox Main Folder/Backend/requirements.txt
deep_translator
SpeechRecognition
flask
flask-cors
pytube
openai-whisper
srt
python-docx
pyPDF2
pytesseract
googletrans
reportlab